[Feat] Modify App CRUD

+++ backend/TmiProject/src/main/java/com/tmi/repository/AppRepository.java
@@ -3,5 +3,5 @@ package com.tmi.repository;
 import com.tmi.entity.App;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppRepository extends JpaRepository<App, Integer> {
+public interface AppRepository extends JpaRepository<App, Long> {
 }
